now we have DEFPROPSPEC, drop idea about DEFPROPLIST defining macros
In fact, there are at least some cases where having DEFPROPSPEC define a macro
would be less flexible than the current approach.  For example, suppose what
we want is for one of the arguments to a property defined with DEFPROPLIST to
be evaluated and then passed as an argument to a property combinator, but
outside of a propapp.  If DEFPROPSPEC defines a macro, then after
macroexpansion this argument will not appear inside a propapp, and so in the
evaluation of DEFPROPLIST's unevaluated propspec, it will not be evaluated.

Signed-off-by: Sean Whitton <spwhitton@spwhitton.name>