rework uploading data to be in terms of CONNECTION-UPLOAD
Incremental updates of items of prerequisite data was not implemented, so the
previous meaning of CONNECTION-UPLOAD was not achieving much.  This simplifies
the core implementation, should still allow connection types to implement the
method to provide optimisations, and provides a way to implement copying items
of prerequisite data into chroots.

Signed-off-by: Sean Whitton <spwhitton@spwhitton.name>