rework bootloader installation so can use property-defining macros
It is better to have the code which installs bootloaders to volumes in
property definitions rather than in plain functions, as then we can specify
that it's :POSIX or :LISP, specify the OS required for it to run, and similar.

This commit enables that by replacing INSTALL-BOOTLOADER with a different
generic which returns propspecs.

Signed-off-by: Sean Whitton <spwhitton@spwhitton.name>