ON-CHANGE: when unapplying, apply the secondary propapps
I believe the new semantics are more useful: when we unapply a property which
has additional properties which will be applied when a change is made, also
apply those properties in just the same way when unapplying the primary
property makes a change.

WITH-UNAPPLY can be used to override, or we might add a variant of ON-CHANGE
which ignores the secondary propapps completely when unapplying.  But I can't
currently envisage a situation in which what is wanted is unapplying the
secondary propapps in reverse order.

Signed-off-by: Sean Whitton <spwhitton@spwhitton.name>