MAKE-PROPSPEC: propagate *CONSFIG* even when propspec expression nil
Previously, MAKE-PROPSPEC would store nil for the SYSTEMS slot in the case
where the PROPSPEC argument was nil.  The idea behind this was that if there
are no properties to apply, then there is no need to load any ASDF systems.
However, the SYSTEMS slot is how a HOST value containing the propspec records
the systems required not just to apply its properties, but also what's
required to instantiate all of its hostattrs.  For example, when some of these
hostattrs are CLOS objects, the classes of those objects must be loaded.  So
don't set SYSTEMS to nil just because the propspec expression is nil.

Also see edbba484918f6de83680a5b29e561fba981e05e9.

Signed-off-by: Sean Whitton <spwhitton@spwhitton.name>