DEFPROPSPEC :DESC/:HOSTATTRS/:CHECK: use WITH-*HOST*-*CONSFIG*
We already evaluate the lambda list inside WITH-*HOST*-*CONSFIG* when
computing the propspec, so this change means we consistently evaluate the
lambda list in that context, rather than evaluating it outside that context in
the case of some property subroutines.

This affects &AUX parameters, &OPTIONAL argument default value forms, etc.,
which would evaluate differently without WITH-*HOST*-CONSFIG*, such as a call
to MAKE-PROPSPEC with no :SYSTEMS keyword parameter supplied.

Signed-off-by: Sean Whitton <spwhitton@spwhitton.name>