MOUNT:UNMOUNTED-BELOW-AND-REMOVED: don't unmount DIR itself
If DIR is itself a mount point then previously we would never delete its
contents because the unmounting is done first.  This meant that when
MOUNT:UNMOUNTED-BELOW-AND-REMOVED was used to remove the root filesystem of a
container or virtual machine, for example, then whether the contents of the
root filesystem was actually deleted depended upon whether DIR happened to be
a mount point.  This change ensures that the deletion is always done.

Signed-off-by: Sean Whitton <spwhitton@spwhitton.name>