rework PUSHNEW-HOSTATTRS to use EQUAL not EQL & add PUSHNEW-HOSTATTR
Previously, PUSHNEW-HOSTATTRS always used EQL to determine whether a hostattr
was already pushed, which would give the wrong answer for many common cases.
Now default to using EQUAL, and provide a way for the caller to specify the
test function.

PUSHNEW-HOSTATTRS now takes a list of hostattrs as a single parameter, rather
than using &REST, in order to make room for the new :TEST keyword parameter.
To mitigate the inconvenience of no longer using &REST, add PUSHNEW-HOSTATTR.

Also fix some parameter and function names in docstrings.

Signed-off-by: Sean Whitton <spwhitton@spwhitton.name>