find XDG_CACHE_HOME without relying on HOME
CONNECTION-CONNATTR for :REMOTE-HOME avoids relying on the HOME environment
variable having the correct value; its docstring describes a case in which
HOME's value can be incorrect.  As determining :XDG-CACHE-HOME depends in most
cases on having the correct value for the remote home directory, don't rely on
the HOME environment variable here either.

Signed-off-by: Sean Whitton <spwhitton@spwhitton.name>