add USER:HAS-ACCOUNT-WITH-UID
The anticipated use case is where both uid and gid are to be set; making GID
an optional keyword argument just allows a shorthand for the case of matching
gid and uid.  The limitation to a Debian-like OS is because of the assumption
of a primary group per user.

Refactor PASSWD-ENTRY to support a new function GROUP-ENTRY that does the same
thing, but with the group database instead of the passwd database.

Signed-off-by: David Bremner <david@tethera.net>