IN-CONSFIG: don't accept NIL as argument
If *CONSFIG* is NIL because the user called (in-consfig nil), then
initialising a propspec without any supplied list of ASDF systems will still
result in a NO-CONSFIG warning, the text of which suggests calling IN-CONSFIG.

For now, avoid that situation by disallowing (in-consfig nil).  If the user
doesn't want to use the *CONSFIG* feature, it is probably best if they muffle
the warning or explicitly supply a list of ASDF systems in each piece of code
that initialises propspecs, rather than declaring (in-consfig nil).

Signed-off-by: Sean Whitton <spwhitton@spwhitton.name>