src/property/file.lisp: unless mode supplied, create missing dirs
This means you can apply properties like FILE:HAS-CONTENT without also having
to explicitly apply FILE:{CONTAINING-,}DIRECTORY-EXISTS or similar.

If missing intermediate directories need to have particular modes or
ownership, it is easy to apply properties to ensure those directories exist
with those attributes before applying the property which will create the file.
Then there are no missing directories for the latter property to create.  In
the case where a mode for the file is supplied, leave it to the caller to
create the directories, as a safety measure (see 487a473390).

In the future we might create missing intermediate directories based on a
supplied mode, such as creating them 0750 when the supplied mode is 0640.

Signed-off-by: Sean Whitton <spwhitton@spwhitton.name>