move OS::ARCHITECTURE slot from OS:LINUX to OS:DEBIANLIKE
We already use this information in the sense of a distribution architecture
rather than a kernel architecture, and the latter are differently grained.
We might later want to have slots for both.

Make OS:SUPPORTS-ARCH-P take OS objects as this simplifies existing usage.

Restore the ORIGINAL-OS parameter to INSTALLER:CLEANLY-INSTALLED-ONCE.  It was
previously replaced in 446b8f4a8ef78cb4605cfb551255bb455be411f0.
CHROOT::%OS-BOOTSTRAPPER-INSTALLED does not need the original host's
architecture in most cases, and accepting an arbitrary propapp means we don't
force the use of PACKAGE:INSTALLED.

Signed-off-by: Sean Whitton <spwhitton@spwhitton.name>