add package CONSFIGURATOR.DATA.UTIL
This package is intended to provide a home for utility functions used by
multiple data sources.  Initially move a local function from
CONSFIGURATOR.DATA.FILES-TREE, and slightly generalize it to support an
extension or TYPE argument.

Note that the goal of LITERAL-DATA-PATHNAME is to map (IDEN1 IDEN2) to
existing paths in a user-maintained file hierarchy.  This is quite different
from DATA-PATHNAME, which escapes various characters to map to a safe internal
filename, effectively flattening a directory hierarchy into a single level.

Signed-off-by: David Bremner <david@tethera.net>