LIBVIRT:DEFINED-FOR: fix two problems with --os-variant
Pass OS values, not HOST values, to LIBVIRT::OS-VARIANT.

Recently virt-install(1) began to error by default if the OS variant is
omitted.  Switch off this requirement as a fallback for versions of Debian
newer than known to virt-install(1) (via libosinfo).

The argument '--osinfo' is recognized by virtinst 3.2.0 and later; in 4.0.0 it
is documented as the preferred form.

Signed-off-by: David Bremner <david@tethera.net>